user_profiler/requirements.txt
# Google ADK & LLMs
google-adk[extensions]>=1.22.1
google-genai>=1.59.0
google-generativeai>=0.8.6
litellm>=1.75.5
mistralai>=1.9.11

# Infrastructure
aiokafka>=0.13.0
qdrant-client>=1.16.2
python-dotenv>=1.2.1
pydantic>=2.12.5
requests